Address review: make name filler generic and fix partial update
- Move LocalizedNamesFiller from Adapter\QuickAccess to Core\Language: it has
  no QuickAccess-specific logic and is reusable (register in core/language.yml)
- Fix partial update data loss: fill() now takes the existing stored values so
  editing a single language no longer overwrites the other languages with the
  auto-fill value; EditQuickAccessHandler passes the loaded names as existing
- Correct the behat scenario accordingly: editing only en-US must preserve the
  stored fr-FR value instead of refilling it
