+++ src/PrestaShopBundle/Resources/config/services/adapter/quick_access.yml
@@ -10,9 +10,6 @@ services:
 
   PrestaShop\PrestaShop\Adapter\QuickAccess\Repository\QuickAccessRepository: ~
 
-  PrestaShop\PrestaShop\Adapter\QuickAccess\LocalizedNamesFiller:
-    autowire: true
-
   PrestaShop\PrestaShop\Adapter\QuickAccess\CommandHandler\AddQuickAccessHandler:
     autowire: true
     autoconfigure: true
